remove navigation and move to different viewset
