+++ crowdsourcing/viewsets/project.py
@@ -142,79 +142,6 @@ ORDER BY relevant_requester_rating desc;
         serializer = ProjectSerializer(instance=projects, many=True, fields=('id', 'name', 'module_count'))
         return Response(serializer.data)
 
-    @list_route(methods=['GET'])
-    def workers_reviews(self, request, **kwargs):
-        projects = request.user.userprofile.requester.project_owner.all()
-        modules = Module.objects.all().filter(project__in=projects)
-        tasks = []
-        module_task_map = {}
-        for module in modules:
-          module_tasks = Task.objects.all().filter(module=module)
-          for tsk in module_tasks:
-            module_task_map[tsk.id] = tsk.module
-          tasks.extend(module_tasks)
-        task_workers = TaskWorker.objects.all().filter(
-            task__in=tasks, task_status__in=[2, 3, 4, 5])
-        serializer = TaskWorkerSerializer(instance=task_workers, many=True)
-        for entry in serializer.data:
-          entry["module"] = module_task_map[entry["task"]].id
-          entry["module_name"] = module_task_map[entry["task"]].name
-          entry["target"] = entry["worker"]
-
-        #dedupe by module and worker
-        pending_reviews = {}
-        for entry in serializer.data:
-          pending_reviews[(entry["module"], entry["worker"])] = entry
-
-        # Get existing ratings
-        ratings = WorkerRequesterRating.objects.all().filter(
-          origin=request.user.userprofile, module__in=modules, type="requester")
-        rating_map = {}
-        for rating in ratings:
-          rating_map[(rating.module.id, rating.target.id)] = rating
-
-        for key, val in rating_map.items():
-          if key in pending_reviews:
-            current_review = pending_reviews[key]
-            current_review["current_rating"] = val.weight
-            current_review["current_rating_id"] = val.id
-
-        return Response(pending_reviews.values())
-
-
-    @list_route(methods=['GET'])
-    def requesters_reviews(self, request, **kwargs):
-        worker = Worker.objects.get(profile=request.user.userprofile)
-        task_workers = TaskWorker.objects.all().filter(worker=worker, task_status__in=[2, 3, 4, 5])
-        modules = []
-        pending_reviews = {}
-
-        for task_worker in task_workers:
-          module = task_worker.task.module
-          modules.append(module)
-          pending_reviews[(module.id, module.project.owner.profile.user.id)] = {
-            "task_worker": TaskWorkerSerializer(instance=task_worker).data,
-            "project": ProjectSerializer(instance=module.project).data,
-            "project_owner_alias": module.project.owner.profile.user.username,
-            "target": module.project.owner.profile.user.id,
-            "module": module.id,
-            "module_data": ModuleSerializer(instance=module).data
-          }
-
-        # Get existing ratings
-        ratings = WorkerRequesterRating.objects.all().filter(
-          origin=worker.profile, module__in=modules, type="worker")
-        rating_map = {}
-        for rating in ratings:
-          rating_map[(rating.module.id, rating.target.id)] = rating
-        for key, val in rating_map.items():
-          if key in pending_reviews:
-            current_review = pending_reviews[key]
-            current_review["current_rating"] = val.weight
-            current_review["current_rating_id"] = val.id
-
-        return Response(pending_reviews.values())
-
 
 
     @list_route(methods=['get'])
