+++ static/templates/base/left_navigation.html
@@ -8,7 +8,7 @@
             <md-list-item ng-class="{'active': isActiveTab('messages')}" ng-click="ctrl.navigateTo('messages')">Messages</md-list-item>
             <md-list-item ng-class="{'active': isActiveTab('task-feed')}" ng-click="ctrl.navigateTo('task-feed')">Task Feed</md-list-item>
             <md-list-item ng-class="{'active': isActiveTab('my-projects')}" ng-click="ctrl.navigateTo('my-projects')">My Projects</md-list-item>
-            <md-list-item ng-class="{'active': isActiveTab('ranking')}" ng-click="ctrl.navigateTo('ranking')">Ranking</md-list-item>
+            <!--<md-list-item ng-class="{'active': isActiveTab('ranking')}" ng-click="ctrl.navigateTo('ranking')">Ranking</md-list-item>-->
             <!-- <md-list-item ng-class="{'active': isActiveTab('bookmarked-projects')}" ng-click="ctrl.navigateTo('bookmarked-projects')">Bookmarked Projects</md-list-item> -->
             <!-- <md-list-item ng-class="{'active': isActiveTab('userskills')}" ng-click="ctrl.navigateTo('userskills')">Skills</md-list-item> -->
             <md-list-item ng-click="ctrl.logout()">Logout</md-list-item>
